now API accept multiple Locations for Origin and destination
