+++ example/directions.dart
@@ -1,9 +1,11 @@
 library google_maps_webservice.directions.example;
 
+import 'dart:io';
+
 import 'package:google_maps_webservice/directions.dart';
 
 final directions =
-    new GoogleMapsDirections(apiKey: 'AIzaSyDW30A0MP3rPs3E6W2K0YRNuCDKQ6kPTco');
+    new GoogleMapsDirections(apiKey: Platform.environment["API_KEY"]);
 
 main() async {
   DirectionsResponse res =
