+++ example/distance.dart
@@ -1,30 +1,36 @@
 library google_maps_webservice.distance.example;
 
+import 'dart:io';
+
 import 'package:google_maps_webservice/distance.dart';
 
 final GoogleDistanceMatrix distanceMatrix =
-    GoogleDistanceMatrix(apiKey: 'AIzaSyDW30A0MP3rPs3E6W2K0YRNuCDKQ6kPTco');
+    GoogleDistanceMatrix(apiKey: Platform.environment["API_KEY"]);
 
 main() async {
-  Location origin = Location(23.721160, 90.394435);
-  Location destination = Location(23.726346, 90.377117);
-
-  String originAddress = 'Bakshibazar,Dhaka';
-  String destinationAddress = 'Banani,Dhaka';
+  List<Location> origins = [
+    Location(23.721160, 90.394435),
+    Location(23.732322, 90.385142),
+  ];
+  List<Location> destinations = [
+    Location(23.726346, 90.377117),
+    Location(23.748519, 90.403121),
+  ];
 
   DistanceResponse responseForLocation =
       await distanceMatrix.distanceWithLocation(
-    origin,
-    destination,
+    origins,
+    destinations,
   );
 
   try {
     print('response ${responseForLocation.status}');
 
     if (responseForLocation.isOkay) {
+      print(responseForLocation.destinationAddress.length);
       responseForLocation.results.forEach((row) {
         row.elements.forEach((element) {
-          print('distance ${element.distance.text}');
+          print('distance ${element.distance.text} duration ${element.duration.text}');
         });
       });
     } else {
@@ -33,6 +39,4 @@ main() async {
   } finally {
     distanceMatrix.dispose();
   }
-
-
 }
