+++ lib/src/directions.dart
@@ -428,23 +428,7 @@ class Step extends _Step {
       : null;
 }
 
-enum TravelMode { driving, walking, bicycling, transit }
 
-TravelMode stringToTravelMode(String mode) {
-  if (mode.toLowerCase() == "driving") return TravelMode.driving;
-  if (mode.toLowerCase() == "walking") return TravelMode.walking;
-  if (mode.toLowerCase() == "bicycling") return TravelMode.bicycling;
-  if (mode.toLowerCase() == "transit") return TravelMode.transit;
-  return null;
-}
-
-String travelModeToString(TravelMode mode) {
-  if (mode == TravelMode.driving) return "driving";
-  if (mode == TravelMode.walking) return "walking";
-  if (mode == TravelMode.bicycling) return "bicycling";
-  if (mode == TravelMode.transit) return "transit";
-  return null;
-}
 
 class Polyline {
   final String points;
@@ -465,90 +449,6 @@ class Value {
       json != null ? new Value(json["value"], json["text"]) : null;
 }
 
-enum RouteType { tolls, highways, ferries, indoor }
-
-RouteType stringToRouteType(String type) {
-  if (type.toLowerCase() == "tolls") return RouteType.tolls;
-  if (type.toLowerCase() == "highways") return RouteType.highways;
-  if (type.toLowerCase() == "ferries") return RouteType.ferries;
-  if (type.toLowerCase() == "indoor") return RouteType.indoor;
-  return null;
-}
-
-String routeTypeToString(RouteType type) {
-  if (type == RouteType.tolls) return "tolls";
-  if (type == RouteType.highways) return "highways";
-  if (type == RouteType.ferries) return "ferries";
-  if (type == RouteType.indoor) return "indoor";
-  return null;
-}
-
-enum Unit { metric, imperial }
-
-Unit stringToUnit(String type) {
-  if (type.toLowerCase() == "metric") return Unit.metric;
-  if (type.toLowerCase() == "imperial") return Unit.imperial;
-  return null;
-}
-
-String unitToString(Unit type) {
-  if (type == Unit.metric) return "metric";
-  if (type == Unit.imperial) return "imperial";
-  return null;
-}
-
-enum TrafficModel { bestGuess, pessimistic, optimistic }
-
-TrafficModel stringToTrafficModel(String type) {
-  if (type.toLowerCase() == "best_guess") return TrafficModel.bestGuess;
-  if (type.toLowerCase() == "pessimistic") return TrafficModel.pessimistic;
-  if (type.toLowerCase() == "optimistic") return TrafficModel.optimistic;
-  return null;
-}
-
-String trafficModelToString(TrafficModel type) {
-  if (type == TrafficModel.bestGuess) return "best_guess";
-  if (type == TrafficModel.pessimistic) return "pessimistic";
-  if (type == TrafficModel.optimistic) return "optimistic";
-  return null;
-}
-
-enum TransitMode { bus, subway, train, tram, rail }
-
-TransitMode stringToTransitMode(String type) {
-  if (type.toLowerCase() == "bus") return TransitMode.bus;
-  if (type.toLowerCase() == "subway") return TransitMode.subway;
-  if (type.toLowerCase() == "train") return TransitMode.train;
-  if (type.toLowerCase() == "tram") return TransitMode.tram;
-  if (type.toLowerCase() == "rail") return TransitMode.rail;
-  return null;
-}
-
-String transitModeToString(TransitMode type) {
-  if (type == TransitMode.bus) return "bus";
-  if (type == TransitMode.subway) return "subway";
-  if (type == TransitMode.train) return "train";
-  if (type == TransitMode.tram) return "tram";
-  if (type == TransitMode.rail) return "rail";
-  return null;
-}
-
-enum TransitRoutingPreferences { lessWalking, fewerTransfers }
-
-TransitRoutingPreferences stringToTransitRoutingPreferences(String type) {
-  if (type.toLowerCase() == "less_walking")
-    return TransitRoutingPreferences.lessWalking;
-  if (type.toLowerCase() == "fewer_transfers")
-    return TransitRoutingPreferences.fewerTransfers;
-  return null;
-}
-
-String transitRoutingPreferencesToString(TransitRoutingPreferences type) {
-  if (type == TransitRoutingPreferences.lessWalking) return "less_walking";
-  if (type == TransitRoutingPreferences.fewerTransfers)
-    return "fewer_transfers";
-  return null;
-}
 
 class Fare extends Value {
   final String currency;
